requirements.txt
httplib2>=0.18.1
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client>=4.1.3
requests>=2.25.1
facebook-sdk>=3.1.0
slacker>=0.14.0
python-dateutil>=2.8.1
tweepy>=3.10.0
pytz>=2020.5

